CI: Clean up pre-release builds (#3040)
Pre-release tests on GitHub actions are taking a long time, almost all
in installation. It turns out the Pandas is uploading sdists, but not a
full set of wheels, to nightly. This adds `--prefer-binary` to just get
the latest wheel they uploaded.

Also, we don't plan to update dependencies during maintenance phases, so
I'm disabling pre-release tests for `maint/` branches. If this is
merged, I'll cherry-pick onto `maint/20.2.x` and `maint/23.1.x`.